Moved benchmark files, streamlined
