--- benchmarks/asv/asv.conf.json
@@ -1,16 +1,0 @@
-{
-    "version": 1,
-    "project": "TransformerEngine",
-    "project_url": "https://github.com/ROCm/TransformerEngine",
-    "repo": "../..",
-    "branches": ["HEAD"],
-    "environment_type": "existing",
-    "install_command": [],
-    "build_command": [],
-    "benchmark_dir": ".",
-    "results_dir": "../.asv/results",
-    "html_dir": "../.asv/html",
-    "install_timeout": 600,
-    "benchmark_timeout": 1200,
-    "launch_method": "spawn"
-}
