--- benchmarks/asv/requirements.txt
@@ -1,3 +1,0 @@
-# Extra dependencies for statistical benchmark comparison
-# (compare_results.py). benchstats pulls in rich, scipy and numpy.
-benchstats>=3.4
