+++ benchmarks/microbenchmarks/asv/requirements.txt
@@ -1,0 +1,4 @@
+# Extra dependencies for statistical benchmark comparison (compare_results.py).
+# benchstats pulls in rich, scipy and numpy.
+numpy
+benchstats>=3.4
